ml: Add D-Bus->Mojo bootstrap with stub Mojo impl
This CL gives a proper implementation of the BootstrapMojoConnection
D-Bus method offered by the ML service daemon.

The caller of that method (Chromium) passes a file descriptor
representing a Unix socket. With this CL, the daemon reads a Mojo
invitation from the socket in order to connect the daemon's Mojo to
Chromium's Mojo process graph. It also extracts a primodial pipe
included by Chromium in the invitation and binds a (placeholder stub)
implementation of the new top-level Mojo interface
MachineLearningService.

The Chromium side of the above interaction is included in
crrev.com/c/1123964.

This CL also brings the .mojom files specifying the ML service Mojo
interface under platform2/ml. Previously these were under
platform/system_api/mojo, with the intent to share them with Chromium,
but Chromium will instead maintain its own eventually-consistent copy of
the mojoms. We are switching to this recommended approach for the
reasons discussed in the following doc:
https://docs.google.com/document/d/1rX3xtR6ZioKZaO-4qBFJFqFypSX1p6gAvQ_MuYn64G8
A separate CL will remove the .mojoms under platform/system_api/mojo.

Testing:

How to automatically test that Chromium + the daemon successfully
perform the D-Bus -> Mojo bootstrap is not clear and is under discussion
(comments welcome). It may have to wait until there is a client in
Chromium calling the daemon, which could then have an end-to-end
autotest.

I manually tested this against a Chromium built with crrev.com/c/1123964
and that also included some test calls to BootstrapMojoConnection
initiated from chrome_browser_main_chromeos.

Scenarios tested:
- Boot (OK for either service to be ready first)
- Sudden restart of UI (daemon should quit itself)
- Sudden restart of daemon (UI should reconnect)

BUG=chromium:836092,chromium:836093
TEST=Manual test per above. How to set up an automated integration test
     for the bootstrap is TBD (future CL).

Change-Id: Ied6309d9a51656dda1af01114b0f05fdb90bfd88
Reviewed-on: https://chromium-review.googlesource.com/1123663
Commit-Ready: Andrew Moylan <amoylan@chromium.org>
Tested-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>