intel: ipu3: Enable YUV reprocessing
The flow for the feature is as below:
1. configure at least 3 streams.
one stream for output YUV,
one stream for input YUV,
one stream for output JPEG.
2. request the output YUV, hal will fill it and return back.
3. the upper will handle the YUV data.
4. request the output JPEG, and the request has one input YUV.
5. hal will encode JPEG by using the input stream as source.
6. then return the output JPEG back to the upper.

If there is input YUV stream, the stream will be used
as the source for the JPEG encoding.

In addition,
1. correct the width and height exif tags in the JPEG.
2. when the main JPEG size is not the same as
the source buffer, it needs to create one YUV internal buffer
to downscale to the JPEG size. correct the wrong
format for the internal buffer.

BUG=b:77833943
TEST=Build and deploy package to nocturne and camera works

Change-Id: Ia18cfc2a9c67ba7dbd49021185e8b51ee31f88b0
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1123721
Commit-Ready: Wei Meng <wei.meng@intel.com>
Tested-by: Wei Meng <wei.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Liang L Yang <liang.l.yang@intel.com>