cryptohome: Implementation for binding LE credentials to PCR
For additional security needed as result of running ARC++ on the
device the low entropy (LE) credentials should pass additional
validation for some PCR values. The PCR indexes and values to be
validated by pinweaver are decided in cryptohome. This CL includes
the required PCR criteria when inserting a new LE credential. Also
it takes care of migration for old credentials that don't have
associated PCR criteria.
The change should work even if the CR50 code is not updated to
validate the PCR values. In this case the call to CR50 is converted
to be sent in the old format, and the response to CheckCredentials
doesn't contain reset_secret.

BUG=chromium:812165
TEST=manual and unit tests
CQ-DEPEND=CL:1165543

Change-Id: Ief724c6c8d46ee1907bca99f361beb5b1dba93b4
Reviewed-on: https://chromium-review.googlesource.com/1124856
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>