intel: ipu3: fix touch AF doesn't work well
[Issue]
touch AF doesn't work well

[Root Cause]
Due to a delay between the AF stats and the actual frame,
the AF is not run on every frame.  Once the bypass is TRUE,
it will use the last afResult (which would be successful from
previous continuous AF).  the trigger is actually being resetted,
and mLastActiveTriggerTime becomes false.  As a result, we will
not do search anymore, as trigger_new_search will never set to TRUE
in the frames afterwards.  In other words, the touch AF command is
being ignored.

[Changes]
should also not do bypass when a AF trigger has been issued

[Test Platform]
nautilus

BUG=b:110856223
TEST=Verified on nautilus that AF can respond to touch commands

Change-Id: Ie0a9da808a1b2cf3b65bdc4756aaae0fd867613a
Signed-off-by: Lai, Jim <jim.lai@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1128578
Commit-Ready: Andy Yeh <andy.yeh@intel.com>
Tested-by: Andy Yeh <andy.yeh@intel.com>
Reviewed-by: Ricky Liang <jcliang@chromium.org>