usb: Use crop and scale to replace stream on/off
The latency for the first camera frame is very long if we do stream on
and off of the camera device. It's about 1 second in nami device.
This CL remove the stream on/off actions. It uses the largest resolution
of the configuration streams and scales the image to different streams.

BUG=b:80284945
TEST=Pass CTS CameraTestCases
  camera_HAL3.device
  camera_HAL3.frame
  camera_HAL3.module
  camera_HAL3.preview
  camera_HAL3.recording
  camera_HAL3.still_capture
  camera_HAL3.stream
  on vayne

Change-Id: I42204deb5f18a1d4238dd50bc3476a2ce595095a
Reviewed-on: https://chromium-review.googlesource.com/1125656
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>