shill: update the bgscan parameters
We identified that shill is very aggressive with the
background scan parameters.
The threshold between short and long interval is very low,
meaning that even when the signal is very strong, we'll fire
a scan very often. Fix this by doing two things:

1) Raise the threshold for short scan internal to -72db
2) Lower the scan frequency when we have a low RSSI to a
   minute (64 seconds to take the approximate scan duration
   into account).

On the way, increase the frequency of the scan when we have
a good RSSI to 6 minutes. Waiting for an hour to issue a
scan is really long.

BUG=b:111071506
TEST=Build and deploy on Eve in Google desk area. Note number of roams
before and after the bgscan parameter change.

Change-Id: If7c57e0375af4a38d3ad5627b10497d243be0068