authpolicy: Back up auth state to user's Cryptohome
Authpolicyd stores the user's authentication state (Kerberos ticket
etc.) in memory. Whenever the user logs out and back in, authpolicyd is
restarted and loses that state. If the user is online during login, the
state is renewed. However, if they are offline during login, the state
cannot be renewed since the Kerberos ticket cannot be requested from the
Active Directory server. Once they go online again, they see a popup
asking to relog since the auth data is not valid. Policy fetch and
Kerberos SSO are not available during such a session and people are
going to have a bad time.

This CL fixes this by backing up user authentication state to the user's
Cryptohome and restoring it whenever needed, e.g. after a restart when
the user logs in while offline.

The tricky part of the CL is to allow the Cryptohome mount to be seen
from authpolicyd's minijail, since authpolicyd is usually started before
Cryptohome is mounted. To make this possible, Cryptohome creates a
folder at /run/daemon-store/authpolicyd and bind-mounts it to itself as
shared mounts. Authpolicyd remounts that folder as slave inside the
mount namespace. Finally, Cryptohome bind-mounts
/home/root/<user_hash>/authpolicyd into that folder, so that the mount
propagates into authpolicyd's mount namespace.

CQ-DEPEND=CL:1165354

BUG=chromium:738433
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy
     Login while online to fetch new Kerberos ticket. Log out, log back
     in while offline (so TGT fetch fails). Reconnect network. You
     should NOT see a notification asking you to relog since the ticket
     expired.

Change-Id: I9d5ef8228ceec71348b7102d840b16e535291607
Reviewed-on: https://chromium-review.googlesource.com/1113917
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>