hal_adapter: add reprocessing effect manager
Add the reprocessing effect manager, which is in charge of managing
reprocessing effects and vendor tags. The reprocessing effects are the
image effects that are applied by reprocessing on frames captured by
camera previously. The manager provides a mechanism for the effects to
initialize and register their processors.

BUG=b:79901482
TEST=GCA works on Soraka

Change-Id: I7b814b3fe707aa65dbd1ad54200520cc5ecdd8e8
Reviewed-on: https://chromium-review.googlesource.com/1132909
Commit-Ready: Hung-yu Wu <hywu@chromium.org>
Tested-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>