Pipe SUPERVISION_TRANSITION info to ARC on boot.
This setting informs ARC that a user has logged in with a different
account type, and needs to have supervision enabled or disabled. This
happens when a child account graduates to regular account, or when some
accounts opt-in to being supervised as a child. This cl is basically
a copy of
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1069031

BUG=chromium:836246
TEST=FEATURES=test emerge-${BOARD} chromeos-login
CQ-DEPEND=CL:1129700

Change-Id: I5719484d464d9cb9bd625b343f00eb02c708f725
Reviewed-on: https://chromium-review.googlesource.com/1134059
Commit-Ready: Patricia Hongo <phongo@google.com>
Tested-by: Yury Khmel <khmel@google.com>
Reviewed-by: Yury Khmel <khmel@google.com>
Reviewed-by: Yury Khmel <khmel@chromium.org>