vm_tools: concierge: Remove useless parent checks
Because we now base 64 encode the passed in file name, there is no way
for it to reference its parent. Remove the checks that pre-date the
base64 encode.

Add a verification of the cryptohome ID, it is a hex number, so make
sure it only contains hex digits.

Change-Id: Ie90812edfdbdd026d006eba2dd6b661f740fcfbc
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1134411
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>