hermes: Add transaction handling
Add three functions that will directly handle interations with the eSIM.

InitiateTransaction will send bytes to the chip, and keep a map of
transaction numbers to callbacks to run once the eSIM responds with a
matching transaction number.

FinalizeTransaction will be the recipient of the received packets from
the eSIM, in particular the bytes received from
OnFileCanReadWithoutBlocking. If the eSIM signals that there are more
bytes, ContinueTransaction will be called to handle the remaining
packets.

BUG=b:80591752
BUG=chromium:847619
TEST=FEATURES=test emerge-cheza hermes

Change-Id: I04c6ce04a14a1b3da6b0618aa4996b8a47bacaff
Reviewed-on: https://chromium-review.googlesource.com/1134584
Commit-Ready: Jacob Rutherford <jruthe@google.com>
Tested-by: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>