authpolicy: Update encoding for DeviceAutoUpdateTimeRestrictions
Update device_policy_encoder so that it accounts for the change from
proto to string of the DeviceAutoUpdateTimeRestrictions policy.

Ignores DeviceUpdateStagingPercentOfFleetPerWeek for now since it's a
repeated int field, which we don't support so far, and it's going to be
changed to a string field, anyway. It needs to be handled in the next
uprev, though.

CQ-DEPEND=CL:1136538

BUG=chromium:852860
TEST=cros_workon_make authpolicy --test

Change-Id: Ie45b76e57212f86eb5f79b88c538e44628d2ce3b
Reviewed-on: https://chromium-review.googlesource.com/1135897
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>