vm_tools: garcon: Refactor PackageKitProxy
I didn't like the way I did this at all before so I wanted to change it
so that it was much easier to extend with new functionality. Now just a
new class is needed in the implementation file with a few methods
overidden. There was way too much code duplication in the prior
implementation.

BUG=None
TEST=Verified on eve install, info, refresh, updates all work

Change-Id: I84f728c41d1e6766248b6a11ab6d898008beb315
Reviewed-on: https://chromium-review.googlesource.com/1137433
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>