authpolicy: Return separate error for constraint violation
Returns a separate error code for a constraint violation while setting
OU for domain join. This happens when an unprivileged user tries to join
a device with a machine name that already exists elsewhere. The error is
use to improve the error message in Chrome.

CQ-DEPEND=CL:1141955

BUG=chromium:864997
TEST=Tested on device

Change-Id: Ibece7e805aaa61feeab398a491203ddd9d1e4fbd
Reviewed-on: https://chromium-review.googlesource.com/1141956
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>