vm_tools: fix SSH key setup under tremplin
Concierge needs to supply the correct SSH keys to set up a container, in
addition to the keys needed to connect to the container. Adjust tremplin's
startup RPC to use the correct keys.

Also refactor the D-Bus call to request SSH keys, which will be shared with
the SSH forwarding setup code in a follow-on CL.

BUG=chromium:863578
TEST=files app works with tremplin
CQ-DEPEND=CL:1142736,CL:1143022

Change-Id: Ie9142028a583007a7b853e98bcc2359998a1beb9
Reviewed-on: https://chromium-review.googlesource.com/1143023
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>