hermes: Add SM-DP+ Implementation
Add basic functionality to the Smdp interface and the SmdpImpl class by
forwarding the information received from the eSIM chip to the SM-DP+
server. The information is encoded as base64 strings and stored in a
JSON dictionary to be written to the http packet that is sent to the
server.

BUG=b:80591752
BUG=chromium:847619
TEST=emerge-cheza hermes

Change-Id: I7c4e82b2a5aa2eaaf397ba0818ecb79935049c60
Reviewed-on: https://chromium-review.googlesource.com/1148595
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>