bluetooth: Set default dispatcher mode based on LE splitter
Now that the LE splitter is configurable via a config file, we can use
the information to decide the default mode of the dispatcher: Use
BlueZ/NewBlue multiplex if and only if the kernel splitter is enabled.
This mode can still be overridden with the --passthrough flag.

BUG=chromium:866704
TEST=Manual test, run btdispatch with /var/lib/bluetooth/newblue:
     1. Does not exist
     2. Exists with content = "1"
     3. Exists with content = "0"
     4. Exists with content = other strings.

Change-Id: I4970d10be0405a2280b21f1344263aa69835ff15
Reviewed-on: https://chromium-review.googlesource.com/1147606
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>