shill: limit daemon size and make it oom-killable
Did as semenzato@ recommended and walked around for a bit. Shill size
tends to be around 30MB-40MB. This change sets the VM size of shill
to be 200MB. It also makes it oom-killable by setting the score
adjustment to -100.

This is motivated by a shill memory leak.

BUG=chromium:858959
TEST=deployed to device, checked /proc/<pid>/{limits,oom_score_adj}

Change-Id: I7da703bf727ed5fbb5209adf6b22b4e3f81dc63b
Reviewed-on: https://chromium-review.googlesource.com/1149148
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>