crash: Move the cleanup and PID file logic to C++
The C++ version does the cleanup at a normal process termination only (i.e.
does not bother signals at all). This approach should be good enough for
autotest, that's the only thing relying on the PID file (thanks to vapier@
for the investigation).

The temporary files in /tmp do not need to be deleted manually since /tmp
is a unique tmpfs provided by minijail for the sandboxed process.

BUG=chromium:391887
TEST=Confirm the followings:
     1) /run/crash_sender.pid is created and deleted
     2) "crash_sender done" message is logged when mock is enabled
     3) crash sending works as before.

Change-Id: Ied360197daf7f643a56e8c38524f2249aaba9c58
Reviewed-on: https://chromium-review.googlesource.com/1149770
Commit-Ready: Satoru Takabayashi <satorux@google.com>
Tested-by: Satoru Takabayashi <satorux@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>