ippusb_manager: Adding a check to GetMessage that bytes have been read
This change fixes an issue which occurred in the event that the
"message_length" which the socket expects to be the first byte on the
socket is 0. If this was the case the function would try to assign a
string of length -1 which resulted in a crash.

Adding a check to ensure that the number of bytes read from the socket
is positive before attempting to assign to the string.

BUG=None
TEST=Tested manually on device

Change-Id: I0b1de8fb330cb0e5f9195ec880124442d01f5967
Reviewed-on: https://chromium-review.googlesource.com/1150735
Commit-Ready: David Valleau <valleau@chromium.org>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>