cros-disks: use MMC_TYPE property to distinguish between MMC and SD
Chrome uses the DeviceMediaType property reported by cros-disks to
determine if a device is USB, SD, CD/DVD, etc. cros-disks checks if a
device is associated with the mmcblk driver to identify if it is a SD
card or not. However, the mmcblk driver supports SD, SDIO, and MMC. For
the Chrome use cases, only SD cards should be tagged with
DeviceMediaType=DEVICE_MEDIA_SD. This CL changes cros-disks to use the
MMC_TYPE property reported by udev, which distinguishes between MMC and
SD, to identify a SD card.

BUG=None
TEST=Run unit tests.
TEST=Verify cros-disks reports DeviceMediaType=DEVICE_MEDIA_SD for a SD
     card, and DeviceMediaType=DEVICE_MEDIA_UNKNOWN for an internal eMMC
     drive.

Change-Id: Ia886eb0eadc19a5e01c111d3c9404000baeeaedf
Reviewed-on: https://chromium-review.googlesource.com/1150855
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>