shill: Add instrumentation to catch SIGTERM crashes
Shill is crashing due to a SIGTERM sent to it by another process. The
crashes have no stack and no common context. This change adds logging to
catch the SIGTERM and print the pid of the process that sent the
SIGTERM. Once this is root caused this CL can be reverted.

BUG=chromium:865442
TEST=kill -SIGTERM <shill's pid>

Change-Id: I3b44ae1cc3b025463ec46ba3afb2cb52fab3b612
Reviewed-on: https://chromium-review.googlesource.com/1227387
Commit-Ready: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>