bluetooth: Add interface to remove exported property.
We need a mechanism to unexport a property when the impersonated
property is no longer valid (which means the property is removed by the
original service). Additionally the interface
CopyPropertyToExportedProperty is renamed to
SyncPropertyToExportedProperty to reflect the new behavior, which does
not only "copy" the value but also "sync" the valid/invalid state.

Also added a public method ExportedInterface::is_exported(), to be used
when detecting whether the same interface has been exported by another
service.

BUG=chromium:866704
TEST=Added unit test in ExportedObjectManagerWrapperTest.SyncProperty

Change-Id: I7a8950a839827cc93254147eb8968c8d0cb4ff1b
Reviewed-on: https://chromium-review.googlesource.com/1152115
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>