bluetooth: Add multiplex logic to btdispatch
This CL adds a basic multiplex logic to btdispatch which works this way:
If there are more than one service exposing the same object at the same
interface, only the first service is considered and the rest of the
services ignored. The priority of the services is determined by the
order at which they are registered via
ObjectManagerInterfaceMultiplexer::RegisterToObjectManager().

This basic multiplex logic serves as a fallback for object/interface
that is not defined specifically. There will be future works configuring
specific multiplex logic for each object/interface/property, and if not
specified the multiplex logic will fallback to this basic logic.

BUG=chromium:866704
TEST=Added unit test to ImpersonationObjectManagerInterfaceTest
CQ-DEPEND=CL:1152115

Change-Id: I1ef5fc99b6114893e33fd8161d566bf63bff65ee
Reviewed-on: https://chromium-review.googlesource.com/1152528
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>