installer: Add utility for OOBE-related recovery crypto
This commit introduces a `cros_oobe_crypto` utility which supports the
libcrypto-backed EC key generation required by OOBE autoconfig during
recovery, with a minimal size footprint (~12kb) in the recovery initramfs.

Its purpose is to generate a prime256v1 key pair, with the private key
PEM-encoded and the public key compressed and outputted as a series of
33 space-separated hexadecimal bytes. This command is used in recovery
before the TPM is locked, where the public key bytes are written to a TPM
space.

More detailed info is given by `cros_oobe_crypto help`.

Generating EC key pairs is easily implemented using the `openssl`
binary, but we do not wish to pull it into the initramfs due to it being
~700kb in size.

BUG=chromium:845589
TEST=generated signatures verify correctly

Change-Id: I8ac3cf9c39c8e5c7b1c1aa2313ee1307216df205
Reviewed-on: https://chromium-review.googlesource.com/1151889
Commit-Ready: Tudor Brindus <tbrindus@chromium.org>
Tested-by: Tudor Brindus <tbrindus@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>