login: Add lsof /drm/dri/card0 output when browser hangs.
We noticed that Chrome fails to authenticate on the drm device
when booting up.
This causes session_manager to kill the browser process that stops
answer to liveness checks.

This CL adds to the output of session manager the list of processes
with /drm/dri/card0 open. In this way we can see if there is any
unexpected process holding the drm device open.

BUG=chromium:865739
TEST=start ui when another process is drm master, then check /var/log/messages

Change-Id: I27bb57e1dc0dcf38993dad8e51d22bdbee766517
Reviewed-on: https://chromium-review.googlesource.com/1152478
Commit-Ready: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>