shill: shill_login_user: Ensure shill user profile dir is owned by shill:shill
We need to chown this dir to be owned by shill:shill even if it already
exists, so that when we update CrOS devices to running shill as
shill:shill (instead of root:root), the dir gets updated to the correct
permissions.

BUG=chromium:867171,chromium:867685,chromium:865184,chromium:867680,chromium:864874,chromium:865789,chromium:866041,chromium:866559,chromium:866961
TEST=flash 10820 image to lulu device. log in as user. configure VPN
stuff. flash new 10888 image to lulu and update shill_login_user script.
login and notice that these problems go away.

Change-Id: Ib487d3a36a7b351887e5c101f1e797a0ab21c3b9
Reviewed-on: https://chromium-review.googlesource.com/1153841
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>