crash: Add crash_sender_paths.h/cc
This is in preparation for moving more paths and functions using the
paths from the shell script. Get() and SetPrefixForTesting() are
introduced to make the path-based functions unit-testable. A unit
test for IsMock() is added based on the test helpers.

Along the way, fix the #include orders.

BUG=chromium:391887
TEST=Run the unit tests

Change-Id: Ibb4194ff9af13458f8dfa0ef1c82b93f0956c795
Reviewed-on: https://chromium-review.googlesource.com/1154785
Commit-Ready: Satoru Takabayashi <satorux@google.com>
Tested-by: Satoru Takabayashi <satorux@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>