crash: treat attached files as relative to the metadata
There's no need for payloads to be in a sep location from the metadata,
and only crash reports are guaranteed to be stable across reboots and
OS upgrades.

For now we normalize the path by hand to support absolute and relative
paths as we might be uploading reports from older OS releases.  We'll
drop support for that after a few releases.

BUG=chromium:866895
TEST=unittests pass

Change-Id: Ibda8c2c24377592d7356364c4acac915722748f8
Reviewed-on: https://chromium-review.googlesource.com/1156064
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>