tools: fix compilation error for generate_camera_profile
Get the number of built-in cameras by enumerating symlinks because the
function |GetCameraDeviceInfos()| is removed in CL:1117974.

This is a temporal workaround fix for snappy, which is the only board
that uses this tool.  We will remove this tool once we migrate it to the
new SKU mapping based approach.

BUG=b:111911591,chromium:868079
TEST=emerge-snappy arc-camera-profile
TEST=Manually run generate_camera_profile on a snappy in the
     lab, and verify its output.

Change-Id: If68e013bc20bae1044696e88f808fdda06dc795f
Reviewed-on: https://chromium-review.googlesource.com/1156341
Commit-Ready: Shik Chen <shik@chromium.org>
Tested-by: Shik Chen <shik@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@chromium.org>