smbprovider: Introduce FakeSambaProxy
This change introdcues FakeSambaProxy. FakeSambaProxy is a proxy
class that has zero logic and passes through all calls to an underlying
FakeSamabaInterface. FakeSambaProxy is needed for testing in order to
return unique_ptrs to SambaInterfaces that are all backed by the same
fake underneath.

BUG=chromium:757625
TEST=unittest
Change-Id: I5dca0745ae9529fd05d91aa9d19490a2e3085626
Reviewed-on: https://chromium-review.googlesource.com/1159081
Commit-Ready: Bailey Berro <baileyberro@chromium.org>
Tested-by: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>