cryptohome: Refactor keyset label logic in HomeDirs
Get rid of duplication of keyset label reading/generation logic.

Besides the pure refactoring, this also slightly changes the
behavior of HomeDirs::GetValidKeyset() to make it consistent with
the rest of the code. Previously, when the label passed to this
method is of form "legacy-NNN", then this method could proceed with
a keyset with the index NNN regardless of whether the keyset has an
explicitly set label. After this change, it would only do that if
the keyset doesn't have the label explicitly set.
It's presumably a safe change, as cryptohome itself always gives
back the explicit label whenever it's present in the keyset, and as
no caller of cryptohome should make up legacy labels on its own.

BUG=none
TEST=existing tests

Change-Id: I83d46708a4bcb37187470d622535dfe11a4d1075
Reviewed-on: https://chromium-review.googlesource.com/1160654
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>