power: Stub out the middleman between SAR sensors and radio controllers
This is a piece of plubming infrastructure that provides a middle-man class
capable of receiving signal from SAR sensors, aggregating them into a vote on
actual user proximity for each class of radio we want to control, and in turn
sending commands to the respective controllers.

As-is, the class is a stub out, and the actual WifiDelegate implementation will
come in a subsequent CL (but see CL:1145860 for further details).

BUG=b:74586744
TEST=added a test case
Signed-off-by: Enrico Granata <egranata@chromium.org>

Change-Id: I8d5cfedd22864dd2d3f13c8e1ffad9aa55e1e3c0
Reviewed-on: https://chromium-review.googlesource.com/1161019
Commit-Ready: Enrico Granata <egranata@chromium.org>
Tested-by: Enrico Granata <egranata@chromium.org>
Reviewed-by: Enrico Granata <egranata@chromium.org>