login: change default oom_score_adj of android container to -100
Android system processes should be killable in case they exhibit memory
leaks.  Without this change, they would have an oom_score_adj of -1000,
making them unkillable. This could lead to a memory leak in a system
application causing a kernel panic. Android application processes will
later have their oom_score_adj furthur increased. This value was chosen
so that as long as a process is using less than 10% of memory, it will
have an oom badness score of 1 (the lowest possible not counting
unkillable processes). The threshold of 10% is an arbitrary line in the
sand.

BUG=b:77539361
TEST=Ensure CtsPermissionTestCases passes
TEST=update unit test

Change-Id: Ifbf2f392f3b1b3ba6d4da521c281a2015be16461
Reviewed-on: https://chromium-review.googlesource.com/1161619
Commit-Ready: Christopher Morin <cmtm@google.com>
Tested-by: Christopher Morin <cmtm@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Vovo Yang <vovoy@chromium.org>