bluetooth: Plumb the pairing state changed event
This makes use of security manager API to register Newblue as the
central pairing observer who is in charge of managing and notifying the
clients. A client need to register itself as the secondary pairing observer
in order to receive the pairing state changed event. As a secondary,
Newblued will received device information along with D-Bus error if any.
There can be multiple secondary pairing observers if, but for now
Newblued is the only client.

This also adds a unit test for pairing state changed events and a unit test
for UniqueID.

BUG=chromium:862849
TEST=run FEATURES=test emerge-$BOARD bluetooth and verify that all unit
     tests pass

Change-Id: I6c1dd09027a90989a0a3d23bf4b1765e951aca63
Reviewed-on: https://chromium-review.googlesource.com/1171211
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>