installer: Fix dd when SRC_BLKSIZE != DST_BLKSIZE
When installing from a USB stick onto a 4k-block device like UFS,
the sector offsets and counts are in different units. It is not
appropriate to just send a bs=<N> argument into dd. Instead, send
ibs= and obs= arguments in.

According to the man page for dd, count= is in input blocks, seek=
is in output blocks, and skip= is in input blocks.

BUG=chromium:832160
TEST=Run chromeos-install on Cheza, with source of USB, dest of UFS.

Change-Id: Idd604bf85a0011d1b84f57e52fcbbca81b91528a
Reviewed-on: https://chromium-review.googlesource.com/1164463
Commit-Ready: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>