hermes: Add DownloadProfile flow to Lpd
On completion of Lpd::Authenticate, the Download procedure can begin.

As specified in SGP.22 3.1.3, this procedure will have the following
flow:

  Esim::PrepareDownload -> Smdp::GetBoundProfilePackage

which will be followed by profile installation.

On the completion of Lpd::Authenticate procedure, Lpd must perform
checks against the Rules Authorization Table of the eSIM chip, as well
as performing checks against the PPR provided by the server, if any,
against the Profiles Info of the eSIM.

BUG=b:80591752
BUG=chromium:847619
TEST=emerge-cheza hermes

Change-Id: Iac30aeefd35f3383e43d9c9e852afb093c2a0d86
Reviewed-on: https://chromium-review.googlesource.com/1171343
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>