bluetooth: Implement method forwarding to multiple services
This adds one rule of method forwarding that the dispatcher needs:
forward the method call to all services. The method call is forwarded
serially to all the services (stop when one of those returns error),
rather than in parallel. This is because in practice we want to minimize
the the unexpected state where one service returns success but the other
returns error.

BUG=chromium:866704
TEST=Added unit test in ImpersionatonObjectManagerInterfaceTest, and
manual test with both BlueZ and NewBlue as the forwarded services:
    bluetoothctl --service org.chromium.Bluetooth
    [bluetooth]# scan on
    [bluetooth]# scan off
During the "scan on", we should see that both BlueZ and NewBlue do
scanning (BREDR and LE, respectively) and we see scan results from both
services.

Change-Id: I80798d46bc99cd293788807403ac7d65a443d511
Reviewed-on: https://chromium-review.googlesource.com/1180612
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>