bluetooth: Rename GetMethodNames to GetMethodForwardings
For future work, we need to configure method forwardings with some kind
of rule, for example "forward to all", or "forward to newblue only".
Therefore the name GetMethodForwardings suits better than GetMethodName
as it configures which methods to be forwarded and how it should be
forwarded.

BUG=chromium:866704
TEST=Unit tests show that there is no behavior change

Change-Id: Ic8c9fa900fa6baf4b15d0d58d6be51f8aa60d2d3
Reviewed-on: https://chromium-review.googlesource.com/1180611
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>