portier: Create a socket for ICMPv6 messages.
This change introduces a socket wrapper for ICMPv6 messages.  Focus is
on the ability to send ICMPv6 messages.  The ability to receive
messages is available, not intended to be used outside of testing.

The 2 types of ICMPv6 messages that are are expected to be generated
by Portier are Destination Unreachable and Packet Too Big error
messages.

TEST=none
BUG=chromium:863260

Change-Id: Ib53ac45040ccbf6434213d1a8db5c8a265cb6ee6
Reviewed-on: https://chromium-review.googlesource.com/1167990
Commit-Ready: Alex Dale <sigquit@google.com>
Tested-by: Alex Dale <sigquit@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>