seneschal: Add storage location field
Add a storage location field to the SharePath message so that it's not
trivially easy for users to share their cookies and other private data
with a VM.  The path included with the message is now treated as being
relative to the storage location and cannot contain ".." components or
end with ".".

Additionally, rename service.proto to seneschal_service.proto.  It turns
out the protobuf compiler only uses the file name when generating the
include guard so it's not possible to include headers from 2 different
files named service.proto even if they are in different packages.

BUG=chromium:703939
TEST=`vmc share termina ../Cookies` doesn't work

Change-Id: I096006dceb1af75c4c80dcec95ed9b8a8129616a
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1166451
Reviewed-by: Dylan Reid <dgreid@chromium.org>