portier: Created proxy interface wrapper.
The proxy interface class is a wrapper around network interfaces that are
to be used in the ND Proxy program.  It initializes several network sockets
to the specification of the neighbor discovery proxy protocol.

The class also handles some of the lower-level ND proxy functions.
This includes:
- Link-layer address substitution in the Neighbor Discovery Message.
- Calculating the ICMPv6 checksums on modified messages.
- Setting of Proxy Bit on Router Advertisements.
- Setting Hop-Limit to 255.

TEST=manual
BUG=chromium:863260

Change-Id: I2f14c9f7a7371c43341da85fbc0d20aa3c857abf
Reviewed-on: https://chromium-review.googlesource.com/1171678
Commit-Ready: Alex Dale <sigquit@google.com>
Tested-by: Alex Dale <sigquit@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>